Moves countrySearch into a file and passes in a callback for cleanliness.
