+++ client/src/app.js
@@ -1,21 +1,4 @@
-
-const onCountriesRequestLoad = function(event) {
-  const resultString = event.target.responseText;
-  let searchResults = JSON.parse(resultString);
-  console.log(searchResults);
-};
-
-const makeCountriesRequest = function(searchString) {
-  const url = `https://restcountries.eu/rest/v2/name/${searchString}`;
-  const request = new XMLHttpRequest();
-  request.open('GET', url);
-  request.addEventListener('load', onCountriesRequestLoad);
-  request.send();
-};
-
-const countrySearch = function(searchString) {
-  makeCountriesRequest(searchString);
-};
+const CountrySearch = require('./countrySearch');
 
 var makeBucketListRequest = function(url, callback){
   var request = new XMLHttpRequest();
@@ -37,7 +20,11 @@ var loadBucketList = function(){
   makeBucketListRequest(url, bucketListRequestComplete);
 };
 
+const renderCountries = function(array) {
+  console.log(array);
+};
+
 window.addEventListener("load", function(){
-  countrySearch('united');
+  new CountrySearch('united', renderCountries);
 	loadBucketList();
 });
